Protect against non-GitHub repositories
I haven't found a way to ban the popups via predicate, but the functions
themselves will now error-out when the current repository's "origin"
remote does not point to GitHub.

It would be possible to use `magit-get-upstream-remote' instead, but hub
isn't smart enough for that yet -- it assumes "origin".

See also github/hub#720.