Make a note about the old nex3/magithub
Fixes #8

I say in the README that this package's name will not be changed -- but
for the record, I may be convinced to do so if there is a *very* good
reason.  Many folks already likely have configuration for this package
at this point, so I'm personally against the idea of breaking their
configurations.