Add an action to submit Magithub bug reports
Hopefully addresses some of the concerns related to #8.  In
nex3/magithub, its issues were often opened up against Magit itself --
this was at least sometimes inappropriate.

Here we add an action to the dispatcher that opens up a new issue on
vermiculus/magithub.  That repository is now (hopefully) set-up to use
the new file /.github/ISSUE_TEMPLATE.md as the issue template, which
should be auto-populated in the issue composition box folks see.