'Crunch' params/data when mocking
We may want to mock up requests with certain parameters.  Of course,
`params' and `data' are full objects and may be far too large to
directly translate into a filename.  (We need this filename -- and the
data inside the file -- to reliably reproduce mocked responses.)

'Crunch' complex data by printing it to a string and taking a hash.
This hash is then tacked onto the filename (generated from the request
itself).

Since params/data are now included in the mock, they're included in
the error for unmocked tests.

This should be improved as needed to somehow recursively sort the data
before crunching it -- just to make things completely realistic.