Drop restriction on cloning while current directory is a repo
It's nonsensical to stop the user from cloning a repo if the clone
default directory is unset and the current directory is a repo, since
whether or not the default directory is set, the user must choose
where to clone the repo.

Instead, trusting the user to pay attention to what they're doing is
fine here.