Update tools for building the manual
Use Ox-Texinfo+ to build the manual.  It looks like you were planning
on doing that anyway but got stuck halfway through when setting up
the build tools.

Instructions on how to use Ox-Texinfo+ can be found at
https://github.com/magit/magit/wiki/Documentation-tools-and-conventions.

The additions to the Makefile and the boilerplate in the manual are
based on what I use for many other packages that come with a proper
manual.  I have left out the parts that are not necessary to build
the manual.

The generated Texi file has to be checked in because Melpa does not
support generating exporting the Org file as Texi.

I have added many FIXMEs to "magithub.org", primarily to mark
incomplete "Key" items (the command is required).