Set up magithub-edit-mode after associating draft with a file
When restoring from a draft (even if the file doesn't exist), setting
the buffer's file clears the major mode for some reason. Setting up
magithub-edit-mode after associating the draft buffer fixes this
problem, and makes magithub edit mode buffers come up in the right
mode regardless of whether there was a draft file associated with them
or not.