Fix CI header nav behavior
Before, we were a bit fragile since the idea of having a collapsible
section in the header wasn't really supported.  With new macro
`magit-insert-section-body', this is made a bit easier :-)

This patch uses the above macro for the section body.  With this same
change, `magit-insert-heading' assumes the content of the actual
header line.